Add system to specify extra results (#39)
I am using NFRT in a non-moddevgradle context, and would like to be able
to get the results of some additional steps out of it. Hardcoding all
the steps someone might want to use doesn't seem practical, so instead I
figured it'd be more sensible to have a system to specify extra steps
whose outputs should be available as results.