Legacy Support (#36)
- NFRT needs the ability to supply a specific JDK home for recompilation
and supplying the JRE libraries for Vineflower / Remapping. (It runs
under 21 itself but the 1.20.1 decompiler crashes with J21 libs)
- Patches in 1.20.1 had a different prefix (not a/, b/), so I added the
required support for specifying the patch prefix. 1.12 specified the
different prefixes in its "userdev" config, so support for that is just
added.
- Properly process the NeoForm/MCP java version in the recompile options
(was harcoded to 21 before)
- NFRT gains support to remap the sources from SRG->Mojang via the
Regular Expression based remapper
- Add a task that generates mappings in different formats and directions
(named <-> intermediary, TSRG/SRG/CSV) for use in pre 1.20.1 toolchains
- Support for a few nitpicky changes in the Userdev JSON

---------

Co-authored-by: Matyrobbrt <matyrobbrt@gmail.com>