Override Vanilla libraries with NeoForm/NeoForge versions when recompiling (#60)
This causes duplicate libraries to be on the classpath, and in the case
of ASM, will cause both ASM 9.7 (from NeoForge) and 9.6 (from Minecraft)
to be acquired.

This PR changes the ExtendedClasspath to work more like the Vanilla
launcher profiles:
When a classpath item that has a Maven coordinate is added, it overrides
any existing item of the same coordinate (ignoring version).