Remove hard-coding of several assumptions about the NeoForm config (#93)
This should allow it to be used with unobf versions of NeoForm.

Behavior for old NeoForm configs should stay unchanged, here's whats
changed in the code:

- The `patch` steps input is no longer hard-wired to the output of the
inject step. The `input` configured in the actual config file is used
instead (for all existing configs this will be `injectOutput`). In new
NeoForm, nothing will be injected.
- Mappings for `splitResources` is now optional. (Although we don't use
it in new NeoForm anyway)
- The `vanillaDeobfuscated` process result is no longer hardwired to
`renameOutput`, instead it will use the input of the decompile step, if
that can be resolved to a node output.